Feat - Adding command flags for the cache and additional for the db
